Added much  improved logging and debugging support by capturing ApacheFOP events and providing them in a response header "ApacheFOP-Serverless-EventLog" if the new "DebugginEnabled" environment configuration value is defined and set to true (default will be false). Also provided easier debugging by supporting a dump of the EventLog if specified via QueryString parameter "&eventLogDump=true" which greatly improves readability when developing with Postman, etc.
Due to the increase in logic, the code is now much better separated into classes so the ApachFOPFunction class is now minimal size again. Also improved project structure. And finally renamed all classes to be consistent with the Apache convention of "Fop" in class names instead of "FOP", but still use the correct name "FOP" in non-class elements like Function Name, etc.