Fixes to address existing XXE XML vulnerabilities now implementing all safeguards recommended by the OWASP XXE Cheatsheet: https://cheatsheetseries.owasp.org/cheatsheets/XML_External_Entity_Prevention_Cheat_Sheet.html#java
We now explicitly opt-in to safe XML processing by enabling secure processing and Block external DTDs & stylesheets and even fully disabling DTDs (doctypes) altogether.
NOTE: Even though XXE vulnerabilities were patched inside some of the FOP internal handling and is now enabled by default this is not the case with any DocuemntBuilderFactory or TransformerFactory which must be instantiate (and are necessary even to run FOP) therefore we must explicitly opt-in (with numerous complex flags) to securely process the XML itself -- it is NOT safe from XXE by default in Java (ugg)!